finish refactoring code
