+++ src/helper.js
@@ -29,4 +29,10 @@ const setDate = () => {
   };
 };
 
-export default setDate;
+const clearElement = (element) => {
+  while (element.firstChild) {
+    element.removeChild(element.firstChild);
+  }
+};
+
+export { setDate, clearElement };
