--- src/unitChange.js
@@ -1,18 +1,0 @@
-
-// let unitWeather = 'metric';
-
-// const switchToCelcius = () => {
-//   const unitC = document.querySelector('.celcius');
-//   unitC.addEventListener('click', () => {
-//     unitWeather = 'metric';
-//   });
-// };
-
-// const switchToFarenheit = () => {
-//   const unitF = document.querySelector('.farenheit');
-//   unitF.addEventListener('click', () => {
-//     unitWeather = 'imperial';
-//   });
-// };
-
-// export { switchToFarenheit, switchToCelcius, unitWeather };
